Move Iceberg data sources to root package
