+++ plugin/trino-iceberg/src/main/java/io/trino/plugin/iceberg/TrinoParquetDataSource.java
@@ -11,13 +11,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.iceberg.io;
+package io.trino.plugin.iceberg;
 
 import io.airlift.slice.Slice;
 import io.trino.parquet.AbstractParquetDataSource;
 import io.trino.parquet.ParquetDataSourceId;
 import io.trino.parquet.ParquetReaderOptions;
 import io.trino.plugin.hive.FileFormatDataSourceStats;
+import io.trino.plugin.iceberg.io.TrinoInput;
+import io.trino.plugin.iceberg.io.TrinoInputFile;
 
 import java.io.IOException;
